requirements.txt
dj-database-url==0.3.0
Django==1.8.4
gunicorn==19.3.0
oauthlib==1.0.3
Pillow==2.9.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.4.0
whitenoise==2.0.4
psycopg2==2.5.4